Remove dependency on Pako
A user reported a while back that Avast was flagging the addon as a virus. I suspect this is due to the usage of Pako.

Since the un-GZIPed JSON is only 1.17MB, we can get away with using the plain JSON in the Firefox build. Since Chromium supports the DecompressionStream API we can still use GZIP there; hopefully Firefox will have support for it soon.