Display body when content OK but NSFW avatar
Instead of doing the full hide when a comment is marked as spam exclusively for its avatar, blur the avatar and username and display an explanation next to the username.

Additionally, un-hide avatar and username on hover.

Hopefully this will resolve complaints by users who still want the content of these comments displayed.

Since this is a significant change, also increment minor version number.